feat: Adds movie detail page and navigation
Introduces a movie detail page to display extended information about each movie.

Adds vue-router for navigation and implements routes for the home page and movie detail page.
The movie item component is updated to include a link to the movie detail page.
The movie star rating component is extracted.
The movie image validation is also updated.
