+++ package.json
@@ -14,7 +14,8 @@
     "@tailwindcss/typography": "^0.5.16",
     "caniuse-lite": "^1.0.30001751",
     "happy-dom": "^20.0.7",
-    "vue": "^3.5.13"
+    "vue": "^3.5.13",
+    "vue-router": "^4.6.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.18.0",
