+++ src/components/MovieItem.vue
@@ -1,22 +1,25 @@
 <script setup lang="ts">
-import { PencilIcon, StarIcon, TrashIcon } from "@heroicons/vue/24/solid";
-import { computed } from "vue";
+import {
+  PencilIcon,
+  StarIcon,
+  TrashIcon,
+  EyeIcon,
+} from "@heroicons/vue/24/solid";
+import MovieStarRating from "@/components/MovieStarRating.vue";
 
 const { movie } = defineProps({ movie: { type: Object, required: true } });
 const emit = defineEmits(["edit", "remove", "update:rating"]);
-
-const movieRated = computed(() => !!movie.rating);
 
 const updateRating = (rating) => {
   emit("update:rating", movie.id, rating);
 };
 
-const removeMovie = () => {
-  emit("remove", movie.id);
-};
-
 const editMovie = () => {
   emit("edit", movie.id);
+};
+
+const removeMovie = () => {
+  emit("remove", movie.id);
 };
 </script>
 
@@ -30,27 +33,7 @@ const editMovie = () => {
         :alt="`${movie.name} poster`"
         class="object-cover object-center h-[600px]"
       />
-      <div class="absolute right-0 top-0 size-16">
-        <StarIcon
-          class="size-16"
-          :class="{
-            'text-yellow-500': movieRated,
-            'text-gray-500': !movieRated,
-          }"
-        />
-        <span
-          v-if="movieRated"
-          class="absolute inset-0 flex items-center justify-center text-xl text-yellow-800"
-        >
-          {{ movie.rating }}
-        </span>
-        <span
-          v-else
-          class="absolute inset-0 flex items-center justify-center text-xl text-gray-400"
-        >
-          -
-        </span>
-      </div>
+      <MovieStarRating :rating="movie.rating" />
     </div>
     <div class="w-full h-56 p-4 flex flex-col items-start justify-start">
       <div class="w-full h-16 shrink-0">
@@ -93,6 +76,12 @@ const editMovie = () => {
           <button class="float-button hover:bg-red-500" @click="removeMovie()">
             <TrashIcon class="size-4" />
           </button>
+          <RouterLink
+            class="float-button hover:bg-red-500"
+            :to="{ name: 'movie-detail', params: { id: movie.id } }"
+          >
+            <EyeIcon class="size-4" />
+          </RouterLink>
         </div>
       </div>
     </div>
